Д.З. 5.14. Модуляция Д.З. 4.16. Исправление замечаний
